refactor(org-domains): extract shared Firecrawl brand extraction
The Firecrawl scrape + branding parse logic was duplicated between
brand-context-extract.ts (MCP tool) and auth.ts (domain-setup
endpoint). Extracted into src/auth/extract-brand.ts with a single
extractBrandFromDomain() function that handles the API call, response
parsing, color/font extraction, and name derivation.

Both callsites now use the shared function, removing ~100 lines of
duplicated code from auth.ts.
