+++ apps/mesh/src/tools/organization/brand-context-extract.ts
@@ -1,93 +1,7 @@
 import { z } from "zod";
 import { defineTool } from "../../core/define-tool";
 import { requireAuth } from "../../core/mesh-context";
-
-/**
- * Map Firecrawl's BrandingProfile to our brand context shape.
- */
-function mapFirecrawlBranding(
-  branding: Record<string, unknown>,
-  metadata: Record<string, unknown>,
-): {
-  logo: string | null;
-  favicon: string | null;
-  ogImage: string | null;
-  fonts: { name: string; role: string }[];
-  colors: { label: string; value: string }[];
-  metadata: Record<string, unknown>;
-} {
-  const images = (branding.images ?? {}) as Record<string, unknown>;
-
-  // Colors: branding.colors is { primary?: string, secondary?: string, ... }
-  const rawColors = (branding.colors ?? {}) as Record<string, unknown>;
-  const colors: { label: string; value: string }[] = [];
-  for (const [label, value] of Object.entries(rawColors)) {
-    if (typeof value === "string" && value) {
-      colors.push({ label, value });
-    }
-  }
-
-  // Fonts: branding.fonts is Array<{ family: string, ... }>
-  // Also check typography.fontFamilies for role-labeled fonts
-  const fonts: { name: string; role: string }[] = [];
-  const typography = (branding.typography ?? {}) as Record<string, unknown>;
-  const fontFamilies = (typography.fontFamilies ?? {}) as Record<
-    string,
-    unknown
-  >;
-
-  // Named font families (primary, heading, code, etc.)
-  const seenFamilies = new Set<string>();
-  for (const [role, family] of Object.entries(fontFamilies)) {
-    if (typeof family === "string" && family) {
-      fonts.push({ name: family, role });
-      seenFamilies.add(family.toLowerCase());
-    }
-  }
-
-  // Additional fonts from the fonts array that weren't in fontFamilies
-  const rawFonts = branding.fonts;
-  if (Array.isArray(rawFonts)) {
-    for (const f of rawFonts) {
-      const family = (f as Record<string, unknown>).family;
-      if (
-        typeof family === "string" &&
-        family &&
-        !seenFamilies.has(family.toLowerCase())
-      ) {
-        fonts.push({ name: family, role: "" });
-        seenFamilies.add(family.toLowerCase());
-      }
-    }
-  }
-
-  // Rich metadata (typography, components, spacing, layout, tone, personality, etc.)
-  const richMetadata: Record<string, unknown> = {};
-  for (const key of [
-    "typography",
-    "components",
-    "spacing",
-    "layout",
-    "animations",
-    "icons",
-    "tone",
-    "personality",
-    "colorScheme",
-  ]) {
-    if (branding[key] !== undefined) {
-      richMetadata[key] = branding[key];
-    }
-  }
-
-  return {
-    logo: (images.logo as string) ?? null,
-    favicon: (images.favicon as string) ?? null,
-    ogImage: (images.ogImage as string) ?? (metadata.ogImage as string) ?? null,
-    fonts,
-    colors,
-    metadata: richMetadata,
-  };
-}
+import { extractBrandFromDomain } from "../../auth/extract-brand";
 
 export const BRAND_CONTEXT_EXTRACT = defineTool({
   name: "BRAND_CONTEXT_EXTRACT",
@@ -132,73 +46,26 @@ export const BRAND_CONTEXT_EXTRACT = defineTool({
       );
     }
 
-    // Normalize domain to URL
-    let url = input.domain.trim();
-    if (!url.startsWith("http")) {
-      url = `https://${url}`;
-    }
-
-    // Call Firecrawl scrape with branding format
-    const response = await fetch("https://api.firecrawl.dev/v1/scrape", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${apiKey}`,
-      },
-      body: JSON.stringify({
-        url,
-        formats: ["branding"],
-      }),
-    });
-
-    if (!response.ok) {
-      const text = await response.text().catch(() => "");
-      throw new Error(
-        `Firecrawl API error: ${response.status} ${text.slice(0, 200)}`,
-      );
-    }
-
-    const result = (await response.json()) as {
-      success?: boolean;
-      data?: {
-        branding?: Record<string, unknown>;
-        metadata?: Record<string, unknown>;
-      };
-    };
-
-    if (!result.success || !result.data?.branding) {
+    const extracted = await extractBrandFromDomain(
+      input.domain,
+      apiKey,
+      input.domain,
+    );
+    if (!extracted) {
       throw new Error("Firecrawl did not return branding data for this URL");
     }
 
-    const branding = result.data.branding;
-    const metadata = result.data.metadata ?? {};
-    const mapped = mapFirecrawlBranding(branding, metadata);
-
-    // Derive name from metadata — prefer the short segment after a separator
-    // in the title (e.g. "Visual CMS for Your Storefront | Deco" → "Deco"),
-    // then ogSiteName, then the domain as last resort.
-    const titleParts = (metadata.title as string)
-      ?.split(/[|–—]/)
-      .map((s) => s.trim())
-      .filter(Boolean);
-    const shortestTitlePart = titleParts
-      ?.slice()
-      .sort((a, b) => a.length - b.length)[0];
-    const name =
-      shortestTitlePart ?? (metadata.ogSiteName as string) ?? input.domain;
-
     const brandData = {
-      name,
-      domain: input.domain,
-      overview: (metadata.description as string) ?? "",
-      logo: mapped.logo,
-      favicon: mapped.favicon,
-      ogImage: mapped.ogImage,
-      fonts: mapped.fonts.length > 0 ? mapped.fonts : null,
-      colors: mapped.colors.length > 0 ? mapped.colors : null,
-      images: null,
-      metadata:
-        Object.keys(mapped.metadata).length > 0 ? mapped.metadata : null,
+      name: extracted.name,
+      domain: extracted.domain,
+      overview: extracted.overview,
+      logo: extracted.logo,
+      favicon: extracted.favicon,
+      ogImage: extracted.ogImage,
+      fonts: extracted.fonts,
+      colors: extracted.colors,
+      images: extracted.images,
+      metadata: extracted.metadata,
     };
 
     // Update existing or create new
